优化x_gen代码结构
