+++ x_gen/x_base/x_base/infinite_vram/offload_manager_save_memory.py
@@ -5,8 +5,10 @@ from collections.abc import Sequence
 import torch
 import torch.distributed as dist
 
+from .offload_common import OffloadCommonMixin
 
-class OffloadManager_For_Save_Memory:
+
+class OffloadManager_For_Save_Memory(OffloadCommonMixin):
     def __init__(
         self,
         root: torch.nn.Module,
@@ -99,25 +101,6 @@ class OffloadManager_For_Save_Memory:
 
         prev_mod._so_freed = True
 
-    def _maybe_free_prev_layer(
-        self, prev_idx: int, keep_n: int, grp: Sequence[torch.nn.Module], params: list[list[torch.Tensor]]
-    ) -> None:
-        if prev_idx < keep_n:
-            return
-        prev_mod = grp[prev_idx]
-        needs_free = getattr(prev_mod, "_so_needs_free", False)
-        already_free = getattr(prev_mod, "_so_freed", False)
-        compute_evt = getattr(prev_mod, "_so_compute_evt", None)
-
-        if not (needs_free and not already_free and compute_evt is not None):
-            return
-
-        with torch.cuda.stream(self.h2d_stream):
-            self.h2d_stream.wait_event(compute_evt)
-            for p in params[prev_idx]:
-                self._release_tensor(p)
-        prev_mod._so_freed = True
-
     # === 新增：把“等待本层预取完成”的逻辑封装 ===
     def _wait_boardcast_if_needed(self, module, keep_n: int) -> None:
         if module.index < keep_n:
@@ -201,28 +184,6 @@ class OffloadManager_For_Save_Memory:
 
         return hook
 
-    def _release_factory(self, tag: str):
-        keep_n = self.keep_n[tag]
-
-        def hook(module, _inp, _out):
-            if module.index < keep_n:
-                return
-
-            # 记录本层计算完成事件，但不立即释放显存
-            # 显存释放将在下一层的prefetch_hook中进行
-            evt = torch.cuda.Event()
-            torch.cuda.current_stream().record_event(evt)
-            module._so_compute_evt = evt
-            module._so_needs_free = True
-            module._so_freed = False
-
-        return hook
-
-    @staticmethod
-    def _iter_tensors(m):
-        yield from m.parameters(recurse=True)
-        yield from m.buffers(recurse=True)
-
     def _setup_parameter_offload(self, tag, grp):
         self.npu_space = []
         self.rank = dist.get_rank()
@@ -275,29 +236,6 @@ class OffloadManager_For_Save_Memory:
             for idx in range(self.keep_n[tag], len(grp)):
                 for p in self.layer_params[tag][idx]:
                     self._release_tensor(p)
-
-    @staticmethod
-    def _release_tensor(p: torch.Tensor):
-        """释放张量的显存"""
-        try:
-            p.data.untyped_storage().resize_(0)
-            p.data.resize_(0)
-        except RuntimeError:
-            p.data = torch.empty(0, dtype=p.dtype, device=p.device)
-        p._released = True
-
-    def _restore_all(self):
-        """恢复所有参数到GPU"""
-        for tag, grp in self.groups.items():
-            for idx, m in enumerate(grp):
-                for p in self.layer_params[tag][idx]:
-                    if p.data.untyped_storage().size() == 0:
-                        restored = torch.empty(p.orig_shape, dtype=p.dtype, device=self.device)
-                        restored.copy_(p.p_cpu, non_blocking=False)
-                        p.data = restored
-                    elif p.data.device != self.device:
-                        p.data = p.data.to(self.device, non_blocking=False)
-                    p._released = False
 
     def _remove_hooks(self):
         """移除所有hooks"""
