--- x_gen/x_diffusers/tests/framework/test_hunyuan.py
@@ -1,254 +1,0 @@
-"""
-Unit tests for x_diffusers.framework.transformer.hunyuan module.
-
-Tests cover:
-- HunyuanVideoAttnProcessor2_0 initialization and methods
-- Attention processing with NPU fusion
-- Transformer configuration
-
-Note: These tests verify standalone logic without importing the actual module.
-No sys.modules mocking is needed as we test mathematical operations and shapes.
-"""
-
-import math
-from unittest.mock import MagicMock
-
-import torch
-
-
-class TestHunyuanVideoAttnProcessor2_0:
-    """Tests for HunyuanVideoAttnProcessor2_0 class."""
-
-    def test_init(self):
-        """Test initialization."""
-        has_sdpa = hasattr(torch.nn.functional, "scaled_dot_product_attention")
-        assert has_sdpa == True  # noqa: E712
-
-
-class TestApplyNorms:
-    """Tests for _apply_norms method."""
-
-    def test_apply_norm_q(self):
-        """Test applying norm to query."""
-        query = torch.randn(1, 24, 100, 64)
-        key = torch.randn(1, 24, 100, 64)  # noqa: F841
-
-        # Simulate Q norm
-        norm_q = torch.nn.LayerNorm(64)
-        query_normed = norm_q(query.transpose(1, 2)).transpose(1, 2)
-
-        assert query_normed.shape == query.shape
-
-    def test_apply_norm_k(self):
-        """Test applying norm to key."""
-        query = torch.randn(1, 24, 100, 64)  # noqa: F841
-        key = torch.randn(1, 24, 100, 64)
-
-        norm_k = torch.nn.LayerNorm(64)
-        key_normed = norm_k(key.transpose(1, 2)).transpose(1, 2)
-
-        assert key_normed.shape == key.shape
-
-
-class TestApplyEncProjAndNorm:
-    """Tests for _apply_enc_proj_and_norm method."""
-
-    def test_encoder_projection(self):
-        """Test encoder QKV projection."""
-        batch_size = 1
-        encoder_seq_len = 256
-        hidden_dim = 64 * 24  # head_dim * num_heads
-
-        encoder_hidden_states = torch.randn(batch_size, encoder_seq_len, hidden_dim)
-
-        # Simulate projections
-        encoder_query = encoder_hidden_states  # add_q_proj
-        encoder_key = encoder_hidden_states  # add_k_proj  # noqa: F841
-        encoder_value = encoder_hidden_states  # add_v_proj  # noqa: F841
-
-        assert encoder_query.shape == (batch_size, encoder_seq_len, hidden_dim)
-
-    def test_encoder_reshape(self):
-        """Test encoder reshape for heads."""
-        batch_size, seq_len, hidden_dim = 1, 256, 1536
-        num_heads = 24
-        head_dim = hidden_dim // num_heads
-
-        encoder_query = torch.randn(batch_size, seq_len, hidden_dim)
-
-        # Reshape to (batch, heads, seq, head_dim)
-        encoder_query = encoder_query.view(batch_size, seq_len, num_heads, head_dim).transpose(1, 2)
-
-        assert encoder_query.shape == (batch_size, num_heads, seq_len, head_dim)
-
-
-class TestAttentionCall:
-    """Tests for attention __call__ method."""
-
-    def test_attention_mask_preparation(self):
-        """Test attention mask preparation for NPU."""
-        attention_mask = torch.zeros(1, 1, 100, 256, dtype=torch.bool)
-
-        # NPU kernel requires Sq == Skv
-        # Repeat mask to match
-        attention_mask = ~attention_mask[0, 0]  # shape: (100, 256)
-        # Expand to (100, 256, 256, 256) for batched attention
-        # (100, 256) -> (100, 1, 1, 256) -> repeat to (100, 256, 256, 256)
-        attention_mask = attention_mask.unsqueeze(1).unsqueeze(2)  # (100, 1, 1, 256)
-        attention_mask = attention_mask.repeat(1, 256, 256, 1)  # (100, 256, 256, 256)
-
-        assert attention_mask.shape == (100, 256, 256, 256)
-
-    def test_qkv_projection(self):
-        """Test QKV projection."""
-        batch_size, seq_len, hidden_dim = 1, 100, 1536
-
-        hidden_states = torch.randn(batch_size, seq_len, hidden_dim)
-
-        query = hidden_states  # to_q
-        key = hidden_states  # to_k  # noqa: F841
-        value = hidden_states  # to_v  # noqa: F841
-
-        assert query.shape == (batch_size, seq_len, hidden_dim)
-
-    def test_qkv_reshape(self):
-        """Test QKV reshape for multi-head."""
-        batch_size, seq_len, hidden_dim = 1, 100, 1536
-        num_heads = 24
-        head_dim = hidden_dim // num_heads
-        attn_heads = num_heads
-
-        query = torch.randn(batch_size, seq_len, hidden_dim)
-        query = query.view(batch_size, seq_len, attn_heads, head_dim).transpose(1, 2)
-
-        assert query.shape == (batch_size, attn_heads, seq_len, head_dim)
-
-
-class TestNPUFusionAttention:
-    """Tests for NPU fusion attention."""
-
-    def test_npu_fusion_attention_params(self):
-        """Test NPU fusion attention parameters."""
-        B, N, S, D = 1, 24, 100, 64
-
-        query = torch.randn(B, N, S, D)  # noqa: F841
-        key = torch.randn(B, N, S, D)  # noqa: F841
-        value = torch.randn(B, N, S, D)  # noqa: F841
-
-        # Parameters for npu_fusion_attention
-        head_num = N
-        input_layout = "BNSD"
-        keep_prob = 1.0  # noqa: F841
-        scale = 1 / math.sqrt(D)
-
-        assert head_num == 24
-        assert input_layout == "BNSD"
-        assert scale == 1 / 8.0
-
-    def test_output_reshape(self):
-        """Test output reshape after attention."""
-        B, N, S, D = 1, 24, 100, 64
-
-        hidden_states = torch.randn(B, N, S, D)
-        hidden_states = hidden_states.transpose(1, 2).flatten(2, 3)
-
-        assert hidden_states.shape == (B, S, N * D)
-
-
-class TestOutputSplit:
-    """Tests for output splitting."""
-
-    def test_output_split_encoder_latent(self):
-        """Test splitting output into encoder and latent parts."""
-        batch_size, seq_len, hidden_dim = 1, 356, 1536  # 256 + 100
-        text_seq_length = 256
-
-        hidden_states = torch.randn(batch_size, seq_len, hidden_dim)
-
-        hidden_states_out, encoder_hidden_states = hidden_states.split(
-            [seq_len - text_seq_length, text_seq_length], dim=1
-        )
-
-        assert hidden_states_out.shape == (batch_size, seq_len - text_seq_length, hidden_dim)
-        assert encoder_hidden_states.shape == (batch_size, text_seq_length, hidden_dim)
-
-
-class TestHunyuanVideoConfig:
-    """Tests for HunyuanVideo transformer configuration."""
-
-    def test_default_config_values(self):
-        """Test default configuration values."""
-        # HunyuanVideo defaults
-        patch_size = 2
-        num_attention_heads = 24
-        attention_head_dim = 64
-        in_channels = 16  # noqa: F841
-        out_channels = 16  # noqa: F841
-        ffn_dim = 3072
-
-        assert patch_size == 2
-        assert num_attention_heads == 24
-        assert attention_head_dim == 64
-        assert ffn_dim == 3072
-
-    def test_rope_max_seq_len(self):
-        """Test RoPE max sequence length."""
-        rope_max_seq_len = 1024
-        assert rope_max_seq_len == 1024
-
-
-class TestAdaLayerNormContinuous:
-    """Tests for AdaLayerNormContinuous."""
-
-    def test_continuous_norm_shape(self):
-        """Test continuous adaptive layer norm."""
-        batch_size, seq_len, dim = 1, 100, 1536
-
-        hidden_states = torch.randn(batch_size, seq_len, dim)
-        timestep_emb = torch.randn(batch_size, dim)
-
-        # Would apply continuous normalization
-        # Just check shapes
-        assert hidden_states.shape == (batch_size, seq_len, dim)
-        assert timestep_emb.shape == (batch_size, dim)
-
-
-class TestParallelManagerIntegration:
-    """Tests for parallel manager integration."""
-
-    def test_parallel_enabled(self):
-        """Test parallel manager enabled."""
-        sp_size = 4
-        heads = 24
-
-        # Check divisibility
-        assert heads % sp_size == 0
-
-    def test_parallel_disabled(self):
-        """Test parallel manager disabled."""
-        sp_size = 1  # noqa: F841
-        heads = 24
-
-        # No division needed
-        attn_heads = heads
-        assert attn_heads == 24
-
-
-class TestEncoderProjModes:
-    """Tests for encoder projection modes."""
-
-    def test_use_encoder_proj_mode_1(self):
-        """Test mode 1: add_q_proj is not None."""
-        add_q_proj = MagicMock()  # Not None
-        encoder_hidden_states = torch.randn(1, 256, 1536)
-
-        use_encoder_proj = add_q_proj is None and encoder_hidden_states is not None
-        assert use_encoder_proj == False  # Mode 1 uses add_q_proj  # noqa: E712
-
-    def test_use_encoder_proj_mode_2(self):
-        """Test mode 2: add_q_proj is None."""
-        add_q_proj = None
-        encoder_hidden_states = torch.randn(1, 256, 1536)
-
-        use_encoder_proj = add_q_proj is None and encoder_hidden_states is not None
-        assert use_encoder_proj == True  # Mode 2 uses to_q for encoder  # noqa: E712
